requirements.txt
pandas~=2.2.3
requests~=2.32.3
streamlit~=1.43.1
nltk~=3.9.1
scikit-learn~=1.6.1
numpy~=2.2.3
beautifulsoup4
openai~=1.65.4
pyyaml~=6.0.2
rouge
matplotlib~=3.10.1
GitPython~=3.1.44
gitdb~=4.0.12
h11~=0.14.0
pillow~=11.1.0
pip~=23.2.1
attrs~=25.1.0
distro~=1.9.0
protobuf~=5.29.3
filelock~=3.18.0
regex~=2024.11.6
scipy~=1.15.2
click~=8.1.8
tqdm~=4.67.1
pyparsing~=3.2.1
pytz~=2025.1
toml~=0.10.2
sniffio~=1.3.1
typing_extensions~=4.12.2
smmap~=5.0.2
httpcore~=1.0.7
idna~=3.10
httpx~=0.28.1
jiter~=0.8.2
charset-normalizer~=3.4.1
setuptools~=77.0.3
threadpoolctl~=3.5.0
torch~=2.6.0
mpmath~=1.3.0
sympy~=1.13.1
altair~=5.5.0
jsonschema~=4.23.0
narwhals~=1.29.1
pyarrow~=19.0.1
Jinja2~=3.1.6
packaging~=24.2
referencing~=0.36.2
MarkupSafe~=3.0.2
joblib~=1.4.2
pydantic~=2.10.6
anyio~=4.8.0
pydantic_core~=2.27.2
fsspec~=2025.3.0
python-dateutil~=2.9.0.post0
pydeck~=0.9.1
certifi~=2025.1.31
tzdata~=2025.1
tornado~=6.4.2
urllib3~=2.3.0
six~=1.17.0
annotated-types~=0.7.0
tenacity~=9.0.0
cachetools~=5.5.2
blinker~=1.9.0
jsonschema-specifications~=2024.10.1
rouge_score~=0.1.2
sentence-transformers~=3.4.1